requirements-dev.txt
pyte>=0.8.2
Pillow>=10.0.0
